+++ Engine/Source/Runtime/Source/Assets/AssetTypes/Textures/Texture.h
@@ -188,7 +188,7 @@ namespace Lumina
 
 
         FTextureResource& GetTextureResource() const { return *TextureResource.get(); }
-        uint8 GetNumMips() const { return TextureResource.get() ? TextureResource->Mips.size() : 0u; }
+        uint8 GetNumMips() const { return TextureResource.get() ? (uint8)TextureResource->Mips.size() : 0u; }
 
         /** (Re)build the GPU image to hold mips [InFirstMip, NumMips) and upload them. The bindless slot is
          *  preserved, so materials that already baked this texture's ResourceID keep sampling it -- they
